+++ packages/opencensus-core/test/test-no-record-span.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {CanonicalCode, CoreTracer, LinkType, MessageEventType} from '../src';
+import {CanonicalCode, CoreTracer, LinkType, MessageEventType, SpanKind} from '../src';
 import {NoRecordRootSpan} from '../src/trace/model/no-record/no-record-root-span';
 import {NoRecordSpan} from '../src/trace/model/no-record/no-record-span';
 
@@ -22,8 +22,9 @@ const tracer = new CoreTracer();
 
 describe('NoRecordSpan()', () => {
   it('do not crash', () => {
-    const root = new NoRecordRootSpan(tracer);
-    const noRecordSpan = new NoRecordSpan(root);
+    const noRecordRootSpan =
+        new NoRecordRootSpan(tracer, 'name', SpanKind.SERVER, 'traceid', '');
+    const noRecordSpan = new NoRecordSpan(noRecordRootSpan);
     noRecordSpan.addAnnotation('MyAnnotation');
     noRecordSpan.addAnnotation('MyAnnotation', {myString: 'bar'});
     noRecordSpan.addAnnotation(