+++ packages/opencensus-core/test/test-tracer.ts
@@ -171,14 +171,47 @@ describe('Tracer', () => {
   });
 
   describe('startRootSpan() with sampler never', () => {
+    const tracer = new CoreTracer();
+    const config = {samplingRate: 0} as TracerConfig;
+
     it('should start the new NoRecordRootSpan instance', () => {
-      const tracer = new CoreTracer();
-      const config = {samplingRate: 0} as TracerConfig;
       tracer.start(config);
       tracer.startRootSpan(options, (rootSpan) => {
         assert.ok(rootSpan instanceof NoRecordRootSpan);
       });
     });
+
+    it('should start the new RootSpan instance when always sampling provided at span level',
+       () => {
+         tracer.start(config);
+         tracer.startRootSpan({name: 'test', samplingRate: 1}, (rootSpan) => {
+           assert.ok(rootSpan);
+           assert.ok(tracer.currentRootSpan instanceof RootSpan);
+           assert.strictEqual(tracer.currentRootSpan, rootSpan);
+         });
+       });
+  });
+
+  describe('startRootSpan() with sampler always', () => {
+    const tracer = new CoreTracer();
+    const config = {samplingRate: 1} as TracerConfig;
+
+    it('should start the new RootSpan instance', () => {
+      tracer.start(config);
+      tracer.startRootSpan(options, (rootSpan) => {
+        assert.ok(rootSpan);
+        assert.ok(tracer.currentRootSpan instanceof RootSpan);
+        assert.strictEqual(tracer.currentRootSpan, rootSpan);
+      });
+    });
+
+    it('should start the new NoRecordRootSpan instance when never sampling provided at span level',
+       () => {
+         tracer.start(config);
+         tracer.startRootSpan({name: 'test', samplingRate: 0}, (rootSpan) => {
+           assert.ok(rootSpan instanceof NoRecordRootSpan);
+         });
+       });
   });
 
   describe('startRootSpan() before start()', () => {
@@ -226,18 +259,20 @@ describe('Tracer', () => {
       });
     });
 
-    it('should create the new RootSpan with no propagation', () => {
-      const tracer = new CoreTracer();
-      tracer.start(defaultConfig);
-      traceOptions.spanContext.options = 0x0;
-      tracer.startRootSpan(traceOptions, (rootSpan) => {
-        assert.ok(rootSpan);
-        assert.strictEqual(rootSpan.name, traceOptions.name);
-        assert.strictEqual(rootSpan.kind, traceOptions.kind);
-        assert.notEqual(rootSpan.traceId, spanContextPropagated.traceId);
-        assert.notEqual(rootSpan.parentSpanId, spanContextPropagated.spanId);
-      });
-    });
+    it('should create the new RootSpan with no propagation (options bit is not set)',
+       () => {
+         const tracer = new CoreTracer();
+         tracer.start(defaultConfig);
+         traceOptions.spanContext.options = 0x0;
+         tracer.startRootSpan(traceOptions, (rootSpan) => {
+           assert.ok(rootSpan);
+           assert.strictEqual(rootSpan.name, traceOptions.name);
+           assert.strictEqual(rootSpan.kind, traceOptions.kind);
+           assert.strictEqual(rootSpan.traceId, spanContextPropagated.traceId);
+           assert.strictEqual(
+               rootSpan.parentSpanId, spanContextPropagated.spanId);
+         });
+       });
 
     it('should create a tracer with default TraceParams when no parameters are specified upon initialisation',
        () => {